# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/regisztracio": {
      "post": {
        "description": "Új felhasználó regisztrálása (publikus végpont).",
        "operationId": "AuthController_regisztracio",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSzemelyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Felhasználó sikeresen regisztrálva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          }
        },
        "summary": "Felhasználó regisztráció",
        "tags": [
          "Authentikáció"
        ]
      }
    },
    "/auth/bejelentkezes": {
      "post": {
        "description": "Bejelentkezés email vagy felhasználónév és jelszó segítségével (publikus végpont).",
        "operationId": "AuthController_bejelentkezes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginSzemelyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sikeres bejelentkezés, JWT token visszaadva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                },
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsIn..."
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Hibás email/felhasználónév vagy jelszó"
          }
        },
        "summary": "Felhasználó bejelentkezés",
        "tags": [
          "Authentikáció"
        ]
      }
    },
    "/auth/tokenfrissites": {
      "get": {
        "description": "Bejelentkezett felhasználó a token lejárta előtt meg tudja azt újítani.",
        "operationId": "AuthController_tokenFrissites",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Sikeres művelet, JWT token visszaadva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                },
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsIn..."
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen token"
          }
        },
        "summary": "JWT frissítése",
        "tags": [
          "Authentikáció"
        ]
      }
    },
    "/auth/jelszomodositas": {
      "post": {
        "description": "Belépési adatokkal azonosított jelszómódosítás.",
        "operationId": "AuthController_jelszoModositas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sikeres jelszómódosítás, Személy visszaadva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen felhasználó"
          }
        },
        "summary": "Jelszó módosítása",
        "tags": [
          "Authentikáció"
        ]
      }
    },
...
    "/esemenyek": {
      "get": {
        "description": "Lekéri az összes eseményt admin jogosultsággal.",
        "operationId": "EsemenyekController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Események sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Esemeny"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Összes esemény lekérdezése",
        "tags": [
          "Események"
        ]
      },
      "post": {
        "description": "Új esemény létrehozása admin jogosultsággal.",
        "operationId": "EsemenyekController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEsemenyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Esemény sikeresen létrehozva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Esemeny"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Új esemény létrehozása",
        "tags": [
          "Események"
        ]
      }
    },
    "/esemenyek/{id}": {
      "get": {
        "description": "Lekéri egy adott esemény adatait az azonosító alapján.",
        "operationId": "EsemenyekController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Esemény sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Esemeny"
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Esemény nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Esemény lekérdezése ID alapján",
        "tags": [
          "Események"
        ]
      },
      "patch": {
        "description": "Meglévő esemény módosítása admin jogosultsággal.",
        "operationId": "EsemenyekController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEsemenyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Esemény sikeresen módosítva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Esemeny"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Esemény nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Esemény módosítása",
        "tags": [
          "Események"
        ]
      },
      "delete": {
        "description": "Meglévő esemény törlése admin jogosultsággal.",
        "operationId": "EsemenyekController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Esemény sikeresen törölve"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Esemény nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Esemény törlése",
        "tags": [
          "Események"
        ]
      }
    },
...
    "/feliratkozasok/esemeny/{id}": {
      "get": {
        "description": "Lekéri egy adott eseményhez tartozó összes személyt.",
        "operationId": "FeliratkozasokController_esemenySzemelyei",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személyek sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feliratkozas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Esemény nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Esemény személyeinek lekérdezése",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS keretrendszerrel készített REST API",
    "description": "REST API személyek, címek, események, feliratkozások, hírek és települések kezelésére",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateCimDto": {
        "type": "object",
        "properties": {
          "cim_tipus": {
            "type": "string",
            "description": "A cím típusa",
            "example": "állandó",
            "enum": [
              "állandó",
              "tartózkodási",
              "levelezési"
            ]
          },
          "cim_iranyitoszam": {
            "type": "string",
            "description": "A cím irányítószáma",
            "example": "1234"
          },
          "cim_varos": {
            "type": "string",
            "description": "A cím városa",
            "example": "Budapest"
          },
          "cim_utca": {
            "type": "string",
            "description": "A cím utcája",
            "example": "Fő utca 1"
          }
        },
        "required": [
          "cim_tipus",
          "cim_iranyitoszam",
          "cim_varos",
          "cim_utca"
        ]
      },
      "RegisterSzemelyDto": {
        "type": "object",
        "properties": {
          "szemely_vezeteknev": {
            "type": "string",
            "description": "A személy vezetékneve",
            "example": "Kovács"
          },
          "szemely_keresztnev": {
            "type": "string",
            "description": "A személy keresztneve",
            "example": "Anna"
          },
          "szemely_szuletesi_datum": {
            "format": "date-time",
            "type": "string",
            "description": "A személy születési dátuma (opcionális)",
            "example": "1990-01-01"
          },
          "szemely_anyja_neve": {
            "type": "string",
            "description": "A személy anyjának neve (opcionális)",
            "example": "Szabó Mária"
          },
          "szemely_vegzettseg": {
            "type": "string",
            "description": "A személy végzettsége",
            "example": "középfok",
            "enum": [
              "nincs",
              "alapfok",
              "középfok",
              "felsőfok"
            ]
          },
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe",
            "example": "kovacs.anna@example.com"
          },
          "szemely_telefon": {
            "type": "string",
            "description": "A személy telefonszáma (opcionális)",
            "example": "+36301234567"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava (legalább 8 karakter, kisbetű, nagybetű és szám)",
            "example": "Titkos123"
          },
          "szemely_cimek": {
            "description": "A személyhez tartozó címek (opcionális)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCimDto"
            }
          }
        },
        "required": [
          "szemely_vezeteknev",
          "szemely_keresztnev",
          "szemely_vegzettseg",
          "szemely_email",
          "szemely_felhasznalonev",
          "szemely_jelszo"
        ]
      },
      "Cim": {
        "type": "object",
        "properties": {
          "cim_id": {
            "type": "number",
            "description": "A cím egyedi azonosítója",
            "example": 1
          },
          "cim_tipus": {
            "type": "string",
            "description": "A cím típusa",
            "example": "állandó",
            "enum": [
              "állandó",
              "tartózkodási",
              "levelezési"
            ]
          },
          "cim_iranyitoszam": {
            "type": "string",
            "description": "A cím irányítószáma",
            "example": "1234"
          },
          "cim_varos": {
            "type": "string",
            "description": "A cím városa",
            "example": "Budapest"
          },
          "cim_utca": {
            "type": "string",
            "description": "A cím utcája",
            "example": "Fő utca 1"
          },
          "cim_szemely_id": {
            "type": "number",
            "description": "A címhez tartozó személy azonosítója",
            "example": 1
          }
        },
        "required": [
          "cim_id",
          "cim_tipus",
          "cim_iranyitoszam",
          "cim_varos",
          "cim_utca",
          "cim_szemely_id"
        ]
      },
      "Esemeny": {
        "type": "object",
        "properties": {
          "esemeny_id": {
            "type": "number",
            "description": "Az esemény egyedi azonosítója",
            "example": 1
          },
          "esemeny_nev": {
            "type": "string",
            "description": "Az esemény neve",
            "example": "Éves konferencia"
          },
          "esemeny_leiras": {
            "type": "string",
            "description": "Az esemény leírása",
            "example": "Éves szakmai konferencia a technológiai újításokról"
          },
          "esemeny_helyszin": {
            "type": "string",
            "description": "Az esemény helyszíne",
            "example": "Budapest, Kongresszusi Központ"
          },
          "esemeny_idopont": {
            "format": "date-time",
            "type": "string",
            "description": "Az esemény időpontja",
            "example": "2025-12-01T10:00:00Z"
          },
          "esemeny_max_fo": {
            "type": "number",
            "description": "Az esemény maximális létszáma",
            "example": 100
          },
          "esemeny_feliratkozasok": {
            "description": "Az eseményhez tartozó feliratkozások",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "esemeny_id",
          "esemeny_nev",
          "esemeny_leiras",
          "esemeny_helyszin",
          "esemeny_idopont",
          "esemeny_max_fo",
          "esemeny_feliratkozasok"
        ]
      },
      "Feliratkozas": {
        "type": "object",
        "properties": {
          "feliratkozas_id": {
            "type": "number",
            "description": "A feliratkozás egyedi azonosítója",
            "example": 1
          },
          "feliratkozas_szemely_id": {
            "type": "number",
            "description": "A feliratkozó személy azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny_id": {
            "type": "number",
            "description": "Az esemény azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny": {
            "description": "A feliratkozás eseménye",
            "allOf": [
              {
                "$ref": "#/components/schemas/Esemeny"
              }
            ]
          }
        },
        "required": [
          "feliratkozas_id",
          "feliratkozas_szemely_id",
          "feliratkozas_esemeny_id",
          "feliratkozas_esemeny"
        ]
      },
      "Szemely": {
        "type": "object",
        "properties": {
          "szemely_id": {
            "type": "number",
            "description": "A személy egyedi azonosítója",
            "example": 1
          },
          "szemely_vezeteknev": {
            "type": "string",
            "description": "A személy vezetékneve",
            "example": "Kovács"
          },
          "szemely_keresztnev": {
            "type": "string",
            "description": "A személy keresztneve",
            "example": "Anna"
          },
          "szemely_szuletesi_datum": {
            "type": "object",
            "description": "A személy születési dátuma (opcionális)",
            "example": "1990-01-01"
          },
          "szemely_anyja_neve": {
            "type": "object",
            "description": "A személy anyjának neve (opcionális)",
            "example": "Szabó Mária"
          },
          "szemely_vegzettseg": {
            "type": "string",
            "description": "A személy végzettsége",
            "example": "középfok",
            "enum": [
              "nincs",
              "alapfok",
              "középfok",
              "felsőfok"
            ]
          },
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe",
            "example": "kovacs.anna@example.com"
          },
          "szemely_telefon": {
            "type": "object",
            "description": "A személy telefonszáma (opcionális)",
            "example": "+36301234567"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava (titkosítva tárolva)",
            "example": "Titkos123"
          },
          "szemely_jogosultsag": {
            "type": "string",
            "description": "A személy jogosultsága",
            "example": "user",
            "enum": [
              "nincs",
              "user",
              "admin"
            ]
          },
          "szemely_torolve": {
            "format": "date-time",
            "type": "string",
            "description": "A személy törlésének időpontja (soft delete)",
            "example": "2025-06-28T13:00:00Z"
          },
          "szemely_cimek": {
            "description": "A személyhez tartozó címek",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cim"
            }
          },
          "szemely_feliratkozasok": {
            "description": "A személyhez tartozó feliratkozások",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feliratkozas"
            }
          }
        },
        "required": [
          "szemely_id",
          "szemely_vezeteknev",
          "szemely_keresztnev",
          "szemely_vegzettseg",
          "szemely_email",
          "szemely_felhasznalonev",
          "szemely_jelszo",
          "szemely_jogosultsag",
          "szemely_cimek",
          "szemely_feliratkozasok"
        ]
      },
      "LoginSzemelyDto": {
        "type": "object",
        "properties": {
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe (opcionális, ha felhasználónév meg van adva)",
            "example": "kovacs.anna@example.com"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve (opcionális, ha email cím meg van adva)",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava",
            "example": "Titkos123"
          }
        },
        "required": [
          "szemely_jelszo"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
      "ModifyPasswordDto": {
        "type": "object",
        "properties": {
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe (opcionális, ha felhasználónév meg van adva)",
            "example": "kovacs.anna@example.com"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve (opcionális, ha email cím meg van adva)",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava",
            "example": "Titkos123"
          },
          "szemely_ujjelszo": {
            "type": "string",
            "description": "A személy új jelszava",
            "example": "TitkosUj123"
          }
        },
        "required": [
          "szemely_jelszo",
          "szemely_ujjelszo"
        ]
      },
...
      "CreateEsemenyDto": {
        "type": "object",
        "properties": {
          "esemeny_nev": {
            "type": "string",
            "description": "Az esemény neve",
            "example": "Éves konferencia"
          },
          "esemeny_leiras": {
            "type": "string",
            "description": "Az esemény leírása",
            "example": "Éves szakmai konferencia a technológiai újításokról"
          },
          "esemeny_helyszin": {
            "type": "string",
            "description": "Az esemény helyszíne",
            "example": "Budapest, Kongresszusi Központ"
          },
          "esemeny_idopont": {
            "type": "string",
            "description": "Az esemény időpontja (ISO 8601 formátumban)",
            "example": "2025-12-01T10:00:00Z"
          },
          "esemeny_max_fo": {
            "type": "number",
            "description": "Az esemény maximális létszáma",
            "example": 100
          }
        },
        "required": [
          "esemeny_nev",
          "esemeny_leiras",
          "esemeny_helyszin",
          "esemeny_idopont",
          "esemeny_max_fo"
        ]
      },
      "UpdateEsemenyDto": {
        "type": "object",
        "properties": {
          "esemeny_nev": {
            "type": "string",
            "description": "Az esemény neve (opcionális)",
            "example": "Éves konferencia"
          },
          "esemeny_leiras": {
            "type": "string",
            "description": "Az esemény leírása (opcionális)",
            "example": "Éves szakmai konferencia a technológiai újításokról"
          },
          "esemeny_helyszin": {
            "type": "string",
            "description": "Az esemény helyszíne (opcionális)",
            "example": "Budapest, Kongresszusi Központ"
          },
          "esemeny_idopont": {
            "type": "string",
            "description": "Az esemény időpontja (ISO 8601 formátumban, opcionális)",
            "example": "2025-12-01T10:00:00Z"
          },
          "esemeny_max_fo": {
            "type": "number",
            "description": "Az esemény maximális létszáma (opcionális)",
            "example": 100
          }
        }
      },
...
    }
  }
}